requirements.txt
Django>=4.2,<5.0
Pillow>=10.0
stripe>=14.4.0
requests>=2.31.0
python-dotenv>=1.0.0
